+++ marconi-mamba/src/Marconi/Api/Routes.hs
@@ -4,8 +4,7 @@
 module Marconi.Api.Routes where
 
 import Data.Set (Set)
-import Marconi.Api.Types (UtxoRowWrapper)
-import Marconi.CardanoAPI (TxOutRef)
+import Marconi.Api.Types (TxOutRef, UtxoRowWrapper)
 import Marconi.JsonRpc.Types (JsonRpc, JsonRpcNotification, RawJsonRpc)
 import Servant.API (Get, NoContent, PlainText, Post, ReqBody, (:<|>), (:>))
