+++ marconi-mamba/src/Marconi/Api/Types.hs
@@ -14,34 +14,32 @@
 
 -- |
 -- This module provides support for writing handlers for JSON-RPC endpoints
-module Marconi.Api.Types where
+module Marconi.Api.Types
+    (
+    module Marconi.Api.Types,
+    module Export,
+    ) where
 
-import Cardano.Api (Address, AddressAny, NetworkId, ShelleyAddr, anyAddressInShelleyBasedEra)
+import Cardano.Api (AddressAny, NetworkId, anyAddressInShelleyBasedEra)
 import Control.Concurrent.QSemN (QSemN)
 import Control.Lens (makeClassy)
 import Data.Aeson (ToJSON (toEncoding, toJSON), defaultOptions, genericToEncoding)
-import Data.List.NonEmpty (NonEmpty)
 import Database.SQLite.Simple (Connection)
 import GHC.Generics (Generic)
 import Network.Wai.Handler.Warp (Settings)
 
-import Marconi.CardanoAPI (CurrentEra)
 import Marconi.Index.Utxo (UtxoRow (UtxoRow))
-
-type CardanoAddress = Address ShelleyAddr
-
--- | Typre represents non empty list of Bech32 compatable addresses"
-type TargetAddresses = NonEmpty CardanoAddress
+import Marconi.Types as Export (CurrentEra, TargetAddresses, TxOutRef)
 
 -- | Type represents http port for JSON-RPC
 type RpcPortNumber = Int
 
 data CliArgs = CliArgs
-  { socket          :: FilePath             -- ^ POSIX socket file to communicate with cardano node
-  , dbPath          :: FilePath             -- ^ filepath to local sqlite for utxo index table
-  , httpPort        :: Maybe Int            -- ^ optional tcp/ip port number for JSON-RPC http server
-  , networkId       :: NetworkId            -- ^ cardano network id
-  , targetAddresses :: TargetAddresses      -- ^ white-space sepparated list of Bech32 Cardano Shelley addresses
+  { socket          :: FilePath        -- ^ POSIX socket file to communicate with cardano node
+  , dbPath          :: FilePath        -- ^ filepath to local sqlite for utxo index table
+  , httpPort        :: Maybe Int       -- ^ optional tcp/ip port number for JSON-RPC http server
+  , networkId       :: NetworkId       -- ^ cardano network id
+  , targetAddresses :: TargetAddresses -- ^ white-space sepparated list of Bech32 Cardano Shelley addresses
   } deriving (Show)
 
 newtype DBConfig = DBConfig {
@@ -50,9 +48,9 @@ newtype DBConfig = DBConfig {
 makeClassy ''DBConfig
 
 data DBQueryEnv = DBQueryEnv
-    { _dbConf         :: DBConfig           -- ^ path to dqlite db
+    { _dbConf         :: DBConfig        -- ^ path to dqlite db
     , _queryQSem      :: QSemN           -- ^ used to serialize addess to sqlite
-    , _queryAddresses :: TargetAddresses    -- ^ user provided addresses to filter
+    , _queryAddresses :: TargetAddresses -- ^ user provided addresses to filter
     }
 makeClassy ''DBQueryEnv
