+++ marconi/src/Marconi/CLI.hs
@@ -2,10 +2,12 @@ module Marconi.CLI where
 
 import Cardano.Api qualified as C
 import Data.ByteString.Char8 qualified as C8
-import Data.List.NonEmpty (NonEmpty, fromList, nub)
+import Data.List.NonEmpty (fromList, nub)
 import Data.Proxy (Proxy (Proxy))
 import Data.Text (pack)
 import Options.Applicative qualified as Opt
+
+import Marconi.Types (TargetAddresses)
 
 chainPointParser :: Opt.Parser C.ChainPoint
 chainPointParser =
@@ -22,11 +24,6 @@ chainPointParser =
       either (const Nothing) Just
       . C.deserialiseFromRawBytesHex (C.proxyToAsType Proxy)
       . C8.pack
-
-type CardanoAddress = C.Address C.ShelleyAddr
-
--- | Typre represents non empty list of Bech32 compatable addresses"
-type TargetAddresses = NonEmpty CardanoAddress
 
 -- | parses a white space separated address list
 -- Note, duplicate addresses are rmoved
